Добавляет переключение фильтров
