+++ js/main.js
@@ -1,13 +1,15 @@
 import {getData} from './api.js';
 import {createThumbnails} from './create-thumbnails.js';
 import {showBigPicture} from './big-picture.js';
-import {setDefaultSlider} from './filters.js';
+import {setDefaultSlider} from './effects.js';
 import {setUserFormSubmit} from './form.js';
 import { showAlert } from './util.js';
+import {showFilters} from './filters.js';
 
 getData().then((picturesArray)=>{
   createThumbnails(picturesArray);
   showBigPicture(picturesArray);
+  showFilters(picturesArray);
 })
   .catch(
     (err) => {
